Add #/bin/bash at the top of bash script
This fixes the second part of Issue 2.
README.md instructs to execute the scipt as
./scripts/init-for-bootstrapping.sh
If the user's shell believes itself to be bash-compatible, it will
use itself to interpret the script, but if it's not quite exactly
correct in its implementation of the construct implemented here to
figure out the paths, the problem arises.  Specifying the interpreter
explicitly helps avoid such problems.