bootstrap: refactor `ObjectMap` hierarchy to use Ring2 model as "definitions"
Each `ObjectMap` sub-instance now contains reference to it's "behavior definition"
(an instance of Ring2's `RGBehavior`). This ahelps to make sure that instances actually
adhere to on-disk class declaration (like, have correct size, and shape, that you cannot set
instance variable that does not exist and so on) and that they're complete (lime, all slots have
defined value).