pst/vm: introduce class `OOP` to hold object references
The new class `OOP` should be used thorough VM code to hold a reference
to an object rather than using raw pointers (i.e., `Object*`). The problem
with raw pointers is that in presence of moving GC, objects held from
native code won't be updated or could be GC'd.

Using somewhat smarter object to hold an object reference allows for
transparent handling of GC-related issues by registering the reference
with GC in ctor and unregistering in dtor.