boostrap: add `MinimalTestSegmentBuilder` for bootstraping test segments
`MinimalTestSegmentBuilder` is a custom segment builder that bootstraps a kernel
and special class containing method `#evaluate` with given source. Instance if this
special class is then set as an entrypoint of the kernel segment, causing runtime to
execute it when segment is loaded.

This can be used for testing the very basics, like:

    MinimalTestSegmentBuilder new
        source: 'evaluate ^0';
        build;
        writeToFile: '../pst/mintest_ret0.bsl'.