pst/vm: use indirect calls to call runtime
This is basically a workaround for lack of sufficient support in OMR
for non-x86 arch. On some architectures, runtime functions may not fit
into branch instruction immediate and thus would require some sort of
trampoline to call. However, trampolines are not implemented in OMR yet.

Using indirect branches always is a way, though they're not implemented
on non-x86 either. However, indirect branches are easier to implement
than trampolines, so we'll use them on RISC-V.